+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Added
 - `vcs.remote_name` that tracks the name of the remote repository (e.g.: `default` on `hg` or `origin` on `git`)
 - `base_ref` parameter that points to the reference (e.g.: a branch) on the base repository
+- `vcs.find_latest_common_revision()` which find the common ancestor between a provided `base_ref` and the current `head_rev`
 
 ## [2.0.0] - 2022-08-01